move html util
