+++ internal/htmlutil/html_test.go
@@ -1,9 +1,9 @@
-package goa4web
+package htmlutil
 
 import "testing"
 
 func TestAnchorLink(t *testing.T) {
-	got := anchorLink("section", "name")
+	got := AnchorLink("section", "name")
 	want := "<a href=\"#section\">name</a><br>"
 	if got != want {
 		t.Errorf("anchorLink=%q", got)
@@ -11,7 +11,7 @@ func TestAnchorLink(t *testing.T) {
 }
 
 func TestPageLink(t *testing.T) {
-	got := pageLink("home", "Home")
+	got := PageLink("home", "Home")
 	want := "<a href=\"?page=home\">Home</a><br>"
 	if got != want {
 		t.Errorf("pageLink=%q", got)
@@ -19,16 +19,16 @@ func TestPageLink(t *testing.T) {
 }
 
 func TestCategoryLevel(t *testing.T) {
-	if got := categoryLevel("cat", 1); got != "<p><a name=\"cat\"><span style=\"font-size: 16;\">cat</span></a><br>\n" {
+	if got := CategoryLevel("cat", 1); got != "<p><a name=\"cat\"><span style=\"font-size: 16;\">cat</span></a><br>\n" {
 		t.Errorf("categoryLevel=%q", got)
 	}
-	if got := categoryLevel("cat", 3); got != "<p><a name=\"cat\"><span style=\"\">cat</span></a><br>\n" {
+	if got := CategoryLevel("cat", 3); got != "<p><a name=\"cat\"><span style=\"\">cat</span></a><br>\n" {
 		t.Errorf("categoryLevel default=%q", got)
 	}
 }
 
 func TestExternalLink(t *testing.T) {
-	got := externalLink("http://x", "X")
+	got := ExternalLink("http://x", "X")
 	want := "<a href=\"http://x\" target=\"_blank\">X</a>"
 	if got != want {
 		t.Errorf("externalLink=%q", got)
@@ -36,7 +36,7 @@ func TestExternalLink(t *testing.T) {
 }
 
 func TestExternalLinkBadURL(t *testing.T) {
-	got := externalLink("javascript:alert(1)", "X")
+	got := ExternalLink("javascript:alert(1)", "X")
 	want := "javascript:alert(1)"
 	if got != want {
 		t.Errorf("externalLink bad=%q", got)
@@ -45,7 +45,7 @@ func TestExternalLinkBadURL(t *testing.T) {
 
 func TestFormatBlobComplex(t *testing.T) {
 	input := "\\</pre\\>line1\n#line2\n-line3\n\\*bold\\* text"
-	got := formatBlob(input)
+	got := FormatBlob(input)
 	want := "</pre>\nline1<br>\n<ol>\n<li>line2</ol>\n<br>\n<ul>\n<li>line3</ul>\n<br>\n<strong>bold</strong> text"
 	if got != want {
 		t.Errorf("unexpected output\n got: %q\nwant: %q", got, want)
@@ -54,7 +54,7 @@ func TestFormatBlobComplex(t *testing.T) {
 
 func TestFormatCategoriesSimple(t *testing.T) {
 	input := "==Cat==\n#item1\n-item2"
-	got := formatCategories(input)
+	got := FormatCategories(input)
 	want := "<ul>\n<a href=\"#Cat\">Cat</a><br><br>\n<br>\n</ul>\n"
 	if got != want {
 		t.Errorf("unexpected output\n got: %q\nwant: %q", got, want)
